+++ finality-aleph/src/network/manager/mod.rs
@@ -1,4 +1,4 @@
-use codec::{Decode, Encode};
+use codec::{Decode, Encode, Error, Input, Output};
 
 use crate::{
     crypto::Signature,
@@ -20,8 +20,6 @@ pub use service::{
     IO as ConnectionIO,
 };
 pub use session::{Handler as SessionHandler, HandlerError as SessionHandlerError};
-
-use crate::compatibility::NetworkDataInSession;
 
 /// Data validators use to authenticate themselves for a single session
 /// and disseminate their addresses.
@@ -50,7 +48,35 @@ impl<M: Multiaddress> AuthData<M> {
 pub type Authentication<M> = (AuthData<M>, Signature);
 
 /// Data inside session, sent to validator network.
-pub type DataInSession<D> = NetworkDataInSession<D>;
+/// Wrapper for data send over network. We need it to ensure compatibility.
+/// The order of the data and session_id is fixed in encode and the decode expects it to be data, session_id.
+/// Since data is versioned, i.e. it's encoding starts with a version number in the standardized way,
+/// this will allow us to retrofit versioning here if we ever need to change this structure.
+#[derive(Clone)]
+pub struct DataInSession<D: Data> {
+    pub data: D,
+    pub session_id: SessionId,
+}
+
+impl<D: Data> Decode for DataInSession<D> {
+    fn decode<I: Input>(input: &mut I) -> Result<Self, Error> {
+        let data = D::decode(input)?;
+        let session_id = SessionId::decode(input)?;
+
+        Ok(Self { data, session_id })
+    }
+}
+
+impl<D: Data> Encode for DataInSession<D> {
+    fn size_hint(&self) -> usize {
+        self.data.size_hint() + self.session_id.size_hint()
+    }
+
+    fn encode_to<T: Output + ?Sized>(&self, dest: &mut T) {
+        self.data.encode_to(dest);
+        self.session_id.encode_to(dest);
+    }
+}
 
 impl<D: Data, M: Multiaddress> From<DataInSession<D>> for NetworkData<D, M> {
     fn from(data: DataInSession<D>) -> Self {
